#248: move embedded elements into separate non-hash versions
