+++ assets/js/elements/person-image.js
@@ -23,7 +23,13 @@ class PersonImage extends LitElement {
   }
 
   render() {
-    var url = `${defaultProfileImage}`;
+    // FIXME: how to best get this value in?
+    let defaultProfileImage = "";
+    if (process.env.DEFAULT_PROFILE_IMAGE != undefined) {
+        defaultProfileImage = `${process.env.DEFAULT_PROFILE_IMAGE}`
+    }
+    // ??
+    var url = `/assets/images/${defaultProfileImage}`;
 
     if (this.thumbnail != "null") {
       url = `${this.thumbnail}`;
