+++ assets/js/vivo-embed-all.js
@@ -1,0 +1,3 @@
+import './vivo-publication-list.js';
+import './vivo-grant-list.js';
+import './vivo-course-list.js';
