+++ templates/docs.html
@@ -88,7 +88,7 @@
         <li><a href="/docs/elements/vivo-tabs">vivo-tabs</a></li>
         <li><a href="/docs/elements/vivo-sidebar">vivo-sidebar</a></li>
         <li><a href="/docs/elements/vivo-truncated-text">vivo-truncated-text</a></li>
-        <li><a href="/docs/elements/test-embed">Embedded Lists</a></li>
+        <li><a href="/embed_demo.html">Embedded Lists</a></li>
       </ul>
     </div>
     <div class="main">
