--- templates/element_doc_pages/test-embed.html
@@ -1,15 +1,0 @@
-
-
-<h1>Testing Embed List</h1>
-<br>
-<h2><u>Embedded Publication List</u></h2>
-<vivo-embedded-pub-list person-url="localhost:4200/entities/person/n5555"></vivo-embedded-pub-list>
-<h2><u>Embedded Grant List</u></h2>
-<vivo-embedded-grant-list person-url="localhost:4200/entities/person/n4449"></vivo-embedded-grant-list>
-<h2><u>Embedded Course List</u></h2>
-<vivo-embedded-course-list person-url="localhost:4200/entities/person/n3323"></vivo-embedded-course-list>
-<h2><u>Embedded Person List for Organization</u></h2>
-<vivo-embedded-organization-people-list type="schools" organization="School of Arts and Sciences"></vivo-embedded-organization-people-list>
-<br>
-<br>
-<br>
